Resolves #1 - Adding contacts.
Contacts can be added from the contacts menu. You cannot add the same contact twice.

Clicking the buttons to the left will take the user to that app with the details filled in.

There has been a change since the original specs, any number can be added as a contact; it does not need to be registered with a user.